make transliteration phrase table thread safe

+++ moses/TranslationModel/PhraseDictionaryTransliteration.cpp
@@ -22,7 +22,7 @@ PhraseDictionaryTransliteration::PhraseDictionaryTransliteration(const std::stri
 
 void PhraseDictionaryTransliteration::CleanUpAfterSentenceProcessing(const InputType& source)
 {
-  RemoveAllInColl(m_allTPColl);
+	ReduceCache();
 }
 
 void PhraseDictionaryTransliteration::GetTargetPhraseCollectionBatch(const InputPathList &inputPathQueue) const
@@ -43,45 +43,67 @@ void PhraseDictionaryTransliteration::GetTargetPhraseCollectionBatch(const Input
     	continue;
     }
 
-	// TRANSLITERATE
-	char *ptr = tmpnam(NULL);
-	string inFile(ptr);
-	ptr = tmpnam(NULL);
-	string outDir(ptr);
-
-	ofstream inStream(inFile.c_str());
-	inStream << sourcePhrase.ToString() << endl;
-	inStream.close();
-
-	string cmd = m_scriptDir + "/Transliteration/prepare-transliteration-phrase-table.pl" +
-			" --transliteration-model-dir " + m_filePath +
-			" --moses-src-dir " + m_mosesDir +
-			" --external-bin-dir " + m_externalDir +
-			" --input-extension " + m_inputLang +
-			" --output-extension " + m_outputLang +
-			" --oov-file " + inFile +
-			" --out-dir " + outDir;
-
-	int ret = system(cmd.c_str());
-	UTIL_THROW_IF2(ret != 0, "Transliteration script error");
-
-	TargetPhraseCollection *tpColl = new TargetPhraseCollection();
-	vector<TargetPhrase*> targetPhrases = CreateTargetPhrases(sourcePhrase, outDir);
-	vector<TargetPhrase*>::const_iterator iter;
-	for (iter = targetPhrases.begin(); iter != targetPhrases.end(); ++iter) {
-		TargetPhrase *tp = *iter;
-		tpColl->Add(tp);
-	}
-
-	m_allTPColl.push_back(tpColl);
-	inputPath.SetTargetPhrases(*this, tpColl, NULL);
-
-	// clean up temporary files
-	remove(inFile.c_str());
-
-	cmd = "rm -rf " + outDir;
-	system(cmd.c_str());
+    GetTargetPhraseCollection(inputPath);
   }
+}
+
+void PhraseDictionaryTransliteration::GetTargetPhraseCollection(InputPath &inputPath) const
+{
+    const Phrase &sourcePhrase = inputPath.GetPhrase();
+    size_t hash = hash_value(sourcePhrase);
+
+    CacheColl &cache = GetCache();
+
+    std::map<size_t, std::pair<const TargetPhraseCollection*, clock_t> >::iterator iter;
+    iter = cache.find(hash);
+
+    if (iter != cache.end()) {
+    	// already in cache
+    	const TargetPhraseCollection *tpColl = iter->second.first;
+    	inputPath.SetTargetPhrases(*this, tpColl, NULL);
+    }
+    else {
+        // TRANSLITERATE
+    	char *ptr = tmpnam(NULL);
+    	string inFile(ptr);
+    	ptr = tmpnam(NULL);
+    	string outDir(ptr);
+
+    	ofstream inStream(inFile.c_str());
+    	inStream << sourcePhrase.ToString() << endl;
+    	inStream.close();
+
+    	string cmd = m_scriptDir + "/Transliteration/prepare-transliteration-phrase-table.pl" +
+    			" --transliteration-model-dir " + m_filePath +
+    			" --moses-src-dir " + m_mosesDir +
+    			" --external-bin-dir " + m_externalDir +
+    			" --input-extension " + m_inputLang +
+    			" --output-extension " + m_outputLang +
+    			" --oov-file " + inFile +
+    			" --out-dir " + outDir;
+
+    	int ret = system(cmd.c_str());
+    	UTIL_THROW_IF2(ret != 0, "Transliteration script error");
+
+    	TargetPhraseCollection *tpColl = new TargetPhraseCollection();
+    	vector<TargetPhrase*> targetPhrases = CreateTargetPhrases(sourcePhrase, outDir);
+    	vector<TargetPhrase*>::const_iterator iter;
+    	for (iter = targetPhrases.begin(); iter != targetPhrases.end(); ++iter) {
+    		TargetPhrase *tp = *iter;
+    		tpColl->Add(tp);
+    	}
+
+    	std::pair<const TargetPhraseCollection*, clock_t> value(tpColl, clock());
+    	cache[hash] = value;
+
+    	inputPath.SetTargetPhrases(*this, tpColl, NULL);
+
+    	// clean up temporary files
+    	remove(inFile.c_str());
+
+    	cmd = "rm -rf " + outDir;
+    	system(cmd.c_str());
+    }
 }
 
 std::vector<TargetPhrase*> PhraseDictionaryTransliteration::CreateTargetPhrases(const Phrase &sourcePhrase, const string &outDir) const
